+++ tests/Logic/XBeteiligingService/XBeteiligungServiceTest.php
@@ -35,7 +35,7 @@ use DemosEurope\DemosplanAddon\XBeteiligung\Logic\XBeteiligungAuditService;
 use DemosEurope\DemosplanAddon\XBeteiligung\Logic\XBeteiligungIncomingMessageParser;
 use DemosEurope\DemosplanAddon\XBeteiligung\Logic\XBeteiligungService;
 use DemosEurope\DemosplanAddon\XBeteiligung\Repository\ProcedureMessageRepository;
-use DemosEurope\DemosplanAddon\XBeteiligung\Repository\XBeteiligungPhaseDefinitionCodeRepository;
+use DemosEurope\DemosplanAddon\XBeteiligung\Repository\XBeteiligungPhaseDefinitionCodeMappingRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
@@ -90,7 +90,7 @@ abstract class XBeteiligungServiceTest extends TestCase
             $reusableMessageBlocks,
             $this->createMock(XBeteiligungAuditService::class),
             new Din91379TextSanitizerService($this->createMock(LoggerInterface::class)),
-            $this->createMock(XBeteiligungPhaseDefinitionCodeRepository::class),
+            $this->createMock(XBeteiligungPhaseDefinitionCodeMappingRepository::class),
         );
     }