+++ webpack.config.js
@@ -4,11 +4,11 @@ const DemosPlanAddon = require('@demos-europe/demosplan-addon-client-builder')
 const config = DemosPlanAddon.build(
   'demosplan-addon-xbeteiligung-async',
   {
-    PhaseCodeFormField: DemosPlanAddon.resolve(
-      'client/hooks/PhaseCreateForm/PhaseCodeFormField.vue'
+    XBeteiligungCreateFormFields: DemosPlanAddon.resolve(
+      'client/hooks/PhaseCreateForm/XBeteiligungCreateFormFields.vue'
     ),
-    PhaseCodeTableField: DemosPlanAddon.resolve(
-      'client/hooks/PhaseListFields/PhaseCodeTableField.vue'
+    XBeteiligungTableFields: DemosPlanAddon.resolve(
+      'client/hooks/PhaseListFields/XBeteiligungTableFields.vue'
     ),
   }
 )